redux, whitespace changes..
