+++ web/source/package.json
@@ -9,12 +9,13 @@
     "@babel/core": "^7.12.13",
     "@babel/preset-env": "^7.12.13",
     "@babel/preset-react": "^7.12.13",
+    "@f0x52/budo-express": "^1.1.0",
+    "@reduxjs/toolkit": "^1.8.5",
     "autoprefixer": "^10.4.8",
     "babelify": "^10.0.0",
     "bluebird": "^3.7.2",
     "browserify": "^17.0.0",
     "browserlist": "^1.0.1",
-    "budo-express": "^1.0.8",
     "css-extract": "^2.0.0",
     "eslint-plugin-react": "^7.24.0",
     "express": "^4.18.1",
@@ -35,7 +36,9 @@
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-error-boundary": "^3.1.4",
+    "react-redux": "^8.0.2",
     "reactify": "^1.1.1",
+    "redux-persist": "^6.0.0",
     "uglifyify": "^5.0.2",
     "wouter": "^2.8.0-alpha.2"
   },
